ENH: first addition of nose testing
Add utility method to the radiomics base class to allow
disabling all features more easily.
Add a test file usable with nosetests that reads in a test
image and mask, calculates the GLCM, then calculates two features.
Added the Matlab features file and the background and segmented volumes from
radiomics-platform as test data.
Added parsing the features file to get the expected baseline values of the features.
Added evaluating the calculated texture features against the baseline. The tests are currently
failing.
Using the template from the GLCM tests, write a first
pass at testing the first order features. Currently, no
mapping between the function names and the baseline so this
just runs the tests, not evaluating the output.