BUG: fix a few bugs in shape features
Running the RadiomicsShape feature calculation from helloRadiomics exposed
a few bugs:
Removed the binWidth argument as it's not used.
Fix the superclass copy/paste error.
padCubicMatrix needed the class name RadiomicsHelpers when calling it, and
np had to be expanded to numpy.
TBD: convert all the getXFeatureValue to only take self as an argument.