ENH: Add Laplacian of Gaussian Feature Class
LoG feature functions are dynamically generated in run-time. This
allows the LoG feature class to maintain the same API as the other
feature classes. NOTE: Some features fail at higher sigma values due to
the increasing homogeneity of the image (binning does not scale with
sigma i.e. filter radius). Exception-handling is sufficient for now, but this
will need to be addressed. Further LoG preprocessing steps and mathematical
implications may also need to be considered.