ENH: add LoG helper
This is a proposed solution to #17, and an alternative to #22 approach proposed
by @vnarayan13.

The idea is to separate steps that take image and produce image into the
RadiomicsHelpers class, and those that take image+mask and produce a list of scalar
features into feature classes.

In the future, provenance should be reconsidered by possibly abstracting each
processing step preceding feature extraction, and maintaining the record together
with the final feature list that would record all steps and corresponding parameters.
That task however requires more careful consideration, and more time. The proposed
solution in the view of the author is more suitable for the current architecture of
pyradiomics.