ENH: updated test utilities
Added and used the verbose flag to clean up output, setting it to off by default.
Added checkResults to the utilities file since it was repeated in all the tests,
updating it to include more checks to avoid dividing by None or 0.
Fixed an erroneous condition check on an index.