ENH: cleaned up and fixed some errors in tests, added docstrings test
Removed the import of csv as that's needed and done in the testUtils.
Cleaned up and checked new utility verbose flag for print outs.
Use the testUtils version of checkResults.
Updated the few tests that were printing docstrings to not do so since those are now
printed and tested in the new test_docstrings file.
Updated test names to reflect feature value names (removed Intensity, Value).
Added imc2 test driver since checkResults can deal with None return value.
In the shape test, return to using the breast1 patient id to match the test names.
Added a test file to check for filled in docstrings on all the feature classes.