BUG: minor tweaks to testing
Use the python logging functionality instead of custom verbose flag. Set the default to only show errors.
Update the test names to remove the _10 as that was the old default bin width.
Change patientID to testCase and only keep it in a variable in the test utilities.
Return a not a number value from a non implemented feature and test for it.