ENH: parmeterize tests over data sets
Update the testUtils file to supply reading and storing the data so that
it's easy to iterate over the data sets.
Added a custom name function to create test names that are alphabetized
in such a way that all the tests for data set 1 are run before recalculating
the features for data set 2, and so on.
Added a utility method to return the set of test cases.

Updated the test files to iterate over data sets. This required using two
global variables, the test utilities and the features. Simplified the imports.

Issue #34