ENH: add LoG to testing
The LoG operation is a preprocessing step to be performed outside the feature classes, so it has to be done in the testing suite while iterating over the range of sigma values.
Updated the testing utilities methods to use a default of 0.0 to signal not applying the LoG.
Only test the 3D LoG results as that's what sitk calculates.

Add the saving, setting, testing of the sigma values to the test suite. For now, it's hard coded to 0.5 to 5.0 in steps of 0.5. In the baseline file, the sigma values are incorporated into the column headers with an underscore instead of a decimal, so added utility methods to convert from floats to strings. In the method to get the baseline feature, construct a different feature name if sigma is set. Iterate over the sigmas for each test case as well as features, so the testing scenarios have three variables.

After discussion with Vivek, he felt that the 1% difference between
baseline Matlab and computed pyradiomics feature values was too strict
and was resulting in too many false failures. With the check set to 1%,
testing the firstorder features and running 825 tests, the result was:
FAILED (errors=54, failures=756)
Setting it to 3%, the result was:
FAILED (errors=54, failures=744)
No change between a setting of 3% and 5%.

Added a graceful failure to the image operations applyLoG so that passing
an invalid sigma results in the return of the unprocessing input image.
Print out an information string rather than causing an exception.

Updated test output to log more clearly when feature names are not found in the baseline.

From comments from @jcfr, use the for ... in construction after verifying
that the lists are iterated in order.
Suggested-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>

Issue #35