ENH: Add API example and testing script
Adapting code from vnarayan13 <narayan.vivek9@gmail.com>:

Added a single variable to adjust when running on a local machine, the location
of the TCGA data set.
Make the output directory if missing, one up from the data directory.

Fix the error that was skipping writing the first patient's values to the output file.
Remove the output files in tests/TCGA-Testing.