add image-resampling
Implement in `radiomics.base`
Check whether `self.interpolator` and `self.resampledpixelspacing` is not none (i.e. no interpolation needed)
add thresholding to the result of the mask interpolation to set all included values in the mask to 1,
as these can have different values due to the interpolation.

Change default resampledpixelspacing to `'3,3,3'`