add image-resampling
Implement in `radiomics.base`
Check whether `self.interpolator` and `self.resampledpixelspacing` is not none (i.e. no interpolation needed)
add thresholding to the result of the mask interpolation to set all included values in the mask to 1,
as these can have different values due to the interpolation.

Change default resampledpixelspacing to `[3,3,3]`

Change function in `imageoperations`;
- add necessary information to `rif`
- calculate size of `resampledImageNode` and pass to `rif`
- add bool `isMask`, default False. When True, applies thresholding to the image to regenerate integer values from interpolated float values.
If this is not done, all interpolated voxels with intensity lower than 1 will revert to 0 (due to cast to type integer)
By applying the threshold, all intensity values  greater than .5 will be set to 1. This has to be done before the image is cast back to type integer, and therefore be implemented in function `interpolateImage`