Change assignment of image and mask in classes derived from `base`
Change assingment form using `inputImage` and `inputMask` to `self.inputImage` and `self.inputMask`, respectively.

During `base.init`, image and mask are read as sitkimage and stored in `self.inputImage` and `self.inputMask`.
However, in the derived classes the image and mask array are generated from `inputImage` and `inputMask`,
which are just the file locations, not the loaded images (so also, no resampling has been applied).