ENH: simplify calculation of `binedges` in `imageoperations.binImage`
Define lowbound as the nearest lower edge of the bin which would contain the minimum value,
where the edges are all equally dividable by `binwidth` (i.e. like when binning is started from 0).
This is accomplished by subtracting minimum modulus binwidth form the minimum value.

Then, only one range needs to be made, using `lowbound` as start, `highbound` as stop and `binwidth` as step.
This produces the same range as generated in commit 11a483c, but does not need 3 seperate ranges.