ENH: split `imageoperations.binImage`
Binning of the image is used in both `glcm.py` and `firstorder.py`
However, in `glcm.py`, it is used to digitize the image, whereas in `firstorder.py`, only the histogram is needed.
This commit splits `binImage` into `getHistogram` and `binImage`. The first bins the image and returns a histogram
(idx 0 = bins with number of voxels they contain, idx 1 = binedges),
no parametermatrix or coordinates are digitized. This is done by `binImage`, which calls to `getHistogram`,
from which it then uses the binedges to digitize the parametermatrix.