ENH: Fix `padVal` checking in constructions of diagonals (rlgl)
When creating the `matrixDiagonals`, a filter is applied to filter out diagonals containing only the `padValue`.
This is to increase performance, as no runlength has to be calculated for these diagonals (as they would
 be excluded during the filling of matrix  `P_rlgl` anyway.
To achieve this, the code uses `numpy.nonzero(x)`, where diagonals containing only zero elements are excluded
(diagonal containing only zero-elements returns no indexes of nonzero elements, therefor the size of the returned array is 0)
However the code checked for elements containing zero in Diags, but `padValue` is set to -2000.
This changes the `numpy.nonzero(x)` to `numpy.nonzero(x != padValue)`, where `x` is the diagonal and
`x != padValue` will evaluate to false (e.g. 0) for the elements containing the `padValue`, and the function
`numpy.nonzero` will return only return the indices of the elements containing a value other than `padValue`