ENH: force datatype of `level` to 'int' in `rlgl.calculateRLGL`
After binning, the values in `self.Matrix` are set to the respective bin the original value belongs to.
These values are used during the filling of `P_rlgl` by subtracting 1 and using them as index in the matrix.
However, these values are stored as 'float64', which produces a warning when used as index.
This adds a cast of `level` to 'int' to prevent this warning and possible errors.