BUG: fix degrees of freedom in calculation of firstorder std and variance
Add `ddof= 1` to `numpy.std`

In the Nature paper, the variance is defined as 1/(N-1) * sum ( (X(i) - uX)**2 ), and std as the sqrt of the variance.
However the standard implementation of `numpy.std` is ( 1/(N) * sum ( (X(i) - uX)**2 ) ) ** 0.5.
To specify whether the divisor should be 'N', 'N-1' or other, one can specify the degrees of freedom in `numpy.std`,
by addin `ddof= 1`, the divisor will be 'N-1'