ENH: sphinx documentation exploration
This is initial work to investigate the suitability and features of sphinx for
documenting feature calculation formulas, as suggested in #53. It seems that this
approach to documentation is a good fit for supporting users of this library, which
are developers by intent.

The process of setting up and using sphinx is summarized in this wiki entry:

https://github.com/Radiomics/pyradiomics/wiki/Documentation-infrastructure

Examples of the documented features are kurtosis and skewness in firstorder.

Sample screenshots will be added to #53.

_build should probably be taken out of the repository, and perhaps _build/html
should be replicated into gh-pages branch, once the repository becomes public.

Some open questions for discussion:
 - should modules.rst and radiomics.rst be included in the repository, or
   generated every time with sphinx-apidoc -f -o . ../radiomics ?

 - documentation for some of the features (texture, wavelets) is quite
   extensive and probably does not fit into this format