STYL: Remove unnecessary variables from `glszm.py`
Remove `self.coefficients['Nr']`, this coeffecient is used in `rlgl.py` to determine maximum size of the RLGL matrix.
However, this is not needed in `glszm.py`, as the maximum size of the matrix is determined by total number of voxels (i.e. `self.coefficients['Np']`).

Remove `n`, remove `labels`, these are filled during calculation of the GLSZM matrix, but not used anywhere.
Remove `B`, this is just a 3x3x3 matrix of ones, and is not used anywhere.
Remove commented-out assignment of `regionSize`, this is not used, and the variable `Y` is not defined.

Reorder `angles` to make it more readable. Some angles are replaced by their complement, but this should not change the outcome. They are still the 13 unique angles possible in a 26-connected region.