ENH: Simplify processing of zone indices
Add code: Remove processed indices from set `ind`
Remove re-calculation of `ind`
Remove implementation of `dataTemp`

Currently, a data matrix (`dataTemp`) is created, which has the same shape as `self.matrix`.
The voxels that have to be processed are set to 1 (initially, all voxels with graylevel i).
From this matrix indices of voxels to be processed are calculated (`ind`), which are processed and then set to zero in `dataTemp`.
After this, `ind` is re-calculated and processed. Repeats until recalculation of `ind` returns an empty list.

This change calculates `ind` directly from `self.matrix`, and then just removes the processed indices from the set, and repeats until `ind` is exhausted.
This removes necessity of creating `dataTemp` and speeds up computation, as the matrix is only checked once per gray level.