BUG: Grow zone from voxel, iterate through zones and gray levels
Add recursive function for growing a zone.
Add generator to also generate 2nd voxel of each angle (by using `for angle_i in angles*-1`)

Currently, only 13 voxels of the 26-connected region of a voxel is assessed for belonging to a certain zone (1 for each angle).
In this case the zone has a maximum size of 14 (center voxel + entire connected region).
However, an angle has 2 voxels associated with it and a size zone matrix also considers the neighbour of a neighbour to belong the same zone.
This commit adds a recursive function which starts with one voxel and gets all neighbouring voxels with an equal gray level.
Then, for all found neighbours, all their unprocessed next-neighbours are assessed for gray level equality, and so on, until no new neighbours are found and the zone is complete.
This recursive function is called until no unprocessed voxels remain, where a set of unprocessed voxels is generated and processed for each gray level.