BUG: Replace recursive function by while loop
During testing, the recursive function produced an error stating that the maximum numbers of recursive calls was exceeded.
Therefore the recursive function has to be replaced. This is done by adding a second while loop, which iterates until region is fully grown.

This is accomplished by defining a stack that is first filled by `ind.pop()` (last element in the set of total elements for a certain gray level).
Then a while loop is started, which pops an element from the stack and checks the nearest neighbours of this element for gray level equality.
All found neighbours are then added to the stack and removed from `ind`. This is loops until stack is exhausted.
Number of iterations in this second while loop is counted, this is equal to the total number of voxels found for the current region and therefore equal to its size.

All remaining indices in `ind` belong to a different region, which will be grown in a next iteration of the first while loop.
Iterates until all regions are grown for current gray level. Repeated for all gray levels.