BUG: Fix calculation of `sumVariance`
Replace `kValuesSum * pxAddy` by the result from sum entropy.
In the Nature publication, the `sumVariance` is calculated as the Sum (for i =2 to 2*ng) of Px+y(i) * ((i - SE)**2), Where SE is the Sum Entropy.
However in the pyradiomics code, the formula subtracts Px+y(i)*i from i, instead of SE.