BUG: Change Log to Log2 in `glcm` entropy coefficients
Change calculation of entropy in `glcm.py` form using logarithm of base e to logarithm of base 2.

In the Nature article, all entropies are calculated using the logarithm of base 2. In the code this is also the case for the calculation of entropy in `firstorder.py`.
However, entropies in `glcm.py` are calculated using the natural logarithm (base e).