ADD: Add variations of `SumVariance` and `SumSquares`
Add `SumVariance2`, this differs from `SumVariance` in that it subtracts `SumAvarage` instead of `SumEntropy` from the intensity before the division.
I this is more in accordance to a variance formula, which is P(i) * (i - mean)^2

Add `SumSquares2`, this version calculates the mean of `P_glcm` differently, as it is not a mean probability, but weighed by (i*j). In variance, the 'i' and the mean are linked in that the mean is the average value of i, however in `SumSquares`, this is not so, as `j` is the intensity, but `u` is the mean of the probability of (i,j).