ENH: Add iteration to reach startleven in `swt3`
Add iteration from 0 to startlevel to generate the correct inputimage for the loop wich generates the decompositions that have to be returned.
Example when startlevel is 1, the data input for the iteration from `start_level` to `start_level + level` will not be the original image, but the first level LLL decomposition of the image.

Additionally, use `numpy.resize(data, shape)` instead of `data.resize(shape)`, this prevents an exception to be thrown