BUG: Add NaN-value handling to testUtils
When a feature produces a NaN-value, it does not add this feature to the dictionary of the testCase. If the situation occurs, where features produce NaN-values in some cases, but  not in others, then an error is thrown in `writeCSV` (missing key).
Use `dict.get()`, this allows specifying a default value, which is returned when key is not found (default value specified as "NaN")