ENH: In case of NaN value, add NaN to self.results
To prevent missing headers when writing a CSV file (which throws an error), values for all feature names need to be added.
In case of a NaN-value, this didn't happen. With the previous commit, a `numpy.nan` value was added to `self.diffs` for these cases.
However, this also needs to happen for `self.results`, as both are written to a  CSV file at the end of testing.