ENH: move cropping/padding outside feature classes
Perform cropping and padding of the tumorvolume at the level of sitkImage, instead of imageArray and maskArray.
This is achieved by using a resampling, where the pixelSpacing remains the same, but the resampling grid is resized to match the cropped tumorvolume (with optional extra padding, specified by padDistance).

Perform cropping and padding in `signatures.py`, instead of in each feature class.
Voxels outside delineation are only set to a padding value inside the feature class, in a call to `imageoperations.getMatrixCoordinates`