ENH: compute shape features on labelMask
For the calculation of shape features the value of gray levels are unimportant, and the shape of the delineation is also captured in the labelmap.
By using the labelmap instead of the original image, it is not necessary to set voxels outside the delineation to a padding value, as these are by definition already 0 in the labelmap (not segmented).

Furthermore, only in shape calculation are voxels outside delineation excluded based upon their intensity value, in other feature classes these voxels are excluded based upon whether or not their indices are a part of `self.matrixCoordinates` (which is based upon the labelmap).
Therefore, by this change, there is no need to set the voxels outside the delineation to a paddingvalue during preprocessing.

 N.B. voxels still need to be set to a paddingvalue in `rlrl.py`, but this is already handled separatly by the code in `rlgl.py`, and is based upon labelmap.