ENH: move cropping/padding outside feature classes Perform cropping and padding of the tumorvolume at the level of sitkImage, instead of imageArray and maskArray. In case where the image is also resampled, this is achieved by using a resampling, where the pixelSpacing remains the same, but the resampling grid is resized to match the cropped tumorvolume (with optional extra padding, specified by padDistance).
In case where there is no resampling necessary, it is achieved by first padding the original image in all directions by padDistance, and then cropping to
a cuboid shape equal to the boundaries of the segementation, increased in all dimensions by padDistance.

Perform cropping and padding in `signatures.py`, instead of in each feature class.
Voxels outside delineation are only set to a padding value inside the feature class, when necessary (in `rlgl.py` and `shape.py`), as there voxels are checked for belonging to the delineation by checking if its value matches the padding Value.

Move assignment of self.imageArray, self.maskArray, self.matrix, self.matrixCoordinates and self.targetVoxelsArray from feature classes to radiomics base.