ENH: Add distance weighted glcm feature class:
Implement new type of glcm feature class. Formulas are the same, but instead of taking the mean of the values calculated for each angle, it calculates the feature on one matrix. This matrix is the sum of the matrices for all angles and distances, where the matrices are weighed for distance to the central voxel.

 W = e^(-||X||^2), where W is the weighting factor, and ||X|| is the norm of the distance. Here, 4 setttings are possible (specified in `kwargs`):
 weightingNorm = None --> No weighting applied (W = 1)
 weightingNorm = infinity --> (Default) Infinity Norm  ( ||X|| = max(|angle|)
 weightingNorm = euclidean --> Euclidean (2nd) Norm  (||X|| = sqrt( sum( (|angle|*pixelspacing)^2 ) ) )
 weightingNorm = manhattan --> Manhattan (1st) Norm  (||X|| = sum( |angle|*pixelspacing ) )