ENH: Add configuration file for Travis CI testing
Private repositories require using http://travis-ci.com, 100 free builds.
Travis doesn't provide support for python projects on Mac.
In order to get it working, we need to use virualenv to install an upgraded
version of pip that will work with a manually downloaded (to avoid
security warnings) specific version of SimpleITK.

TBD: if pyenv is needed to ensure that python 2.7 is installed
for future proofing.

Issue #101