ENH: Switch definition of enabling featureClasses and functions
Settings are stored in self.enabledFeatures. This is a dictionary with featureClassName as key, and as value a list of featureNames (strings).
All classes contained in the key are calculated, if the value is an empty list, all features are enabled. If the list is not empty, only the features present in the list will be calculated for this class. To disable an entire class, remove the key from the dictionary