ENH: Remove support for padding from imageoperations
Padding is removed as a possible setting, as values outside delineation are not used in the pyradiomics code.
Exception is in shape.py, where a padding of 1 is needed to prevent an index-out-of-range exception.
This occurs as for each voxel in the segmentation, it's neighbours are checked for membership in the segmentation. If a neighbouring voxel is not a member, this vertex is counted as being part of the surface area.
If a voxel on the edge of the cropped labelmap image were to be part of the segmentation, the code then tries to assess a voxel outside the image, and so tries to assess a voxel for which there is no valid index in self.maskArray.