ENH: Improve integrated resampling and cropping
Ensure that the resamplegrid of the cropped volume is aligned with the resampledgrid of the the original volume.
This is achieved by calculating the indices of the lower and upper bounds of the bounding box in terms of the resampled coordinate system.
The index for the lowerbound is rounded down and for the upper bound is rounded up to prevent data loss.
Additionally add padding for extra safty against data loss (mostly necessary when upsampling).
The index for the lowerbound is then brought back into the original Index space as a continuous index and is added to the continuous index of the new origin, this ensures that the resampling grid is aligned. This alignment is necessary to ensure that the outcome of the resampling is not influenced by the location of the bounding box.

Rename `cropTumorMaskToCube` to  `croptToTumorMask` to address any confusion caused by use of `cube`