ENH: Add dependency specification to `setup.py`, version `circle.yml`
Add dependencies necessary to run pyradiomics to `setup.py`, this ensures they are correctly installed when pyradiomics package is installed.
This also includes versioning, ensuring the pyradiomics code uses the versions of packages on which it was tested.

Add the versioning to `circle.yml`. This enables the versioning of the packages for the testing.