ENH: Update Baseline
Revert MatlabBaselineFeatures.csv to values of matlab (only changes matlab baseline for sumSquares).

Add new baseline for sumSquares to AdditionalBaseline.csv.
The change in this baseline is due to the fact that the calculation of sumSquares is dependent on the direction of each 3D angle, which is changed due to the dynamic generation of angles. This dependency in turn is caused by the fact that currently, the generate GLCM is not symmetrical, and the direction of the respective angles change the distribution of i and j (transposed for certain angles).