ENH: Add image filters
Square image filter calculates the square of the intensity value, and then normalizes all values back to the original range.
SquareRoot image filter calculates the square root of the absolute intensity value, and then normalizes all values back to the original image
Logarithm image filter calculates the logarithm of the absolute intensity value + 1, and then normalizes all values back to the original range.
Exponential image filter calculates the exponential of the intensity value, normalized to the original range of values.