BUG: Fix for maximum3Ddiameter and partial fix for surface area in shape
maximum3Ddiameter: In a numpy.sum, the axis was missing, causing it to sum all squares in the matrix, instead of just the squares defining the particular euclidean distance.
Additionally, it calculated the distance between voxels as highbound-lowbound + 1, however, this is the number of voxels, not the distance between them. Remove `+1`.

surface area:
simplify counting of faces by generating sums of absolute difference matrices (`surf_x`, `surf_y` and `surf_z`). These count the number of times a neighbouring voxel is different in a certain direction (x, y and z, respectively). Then multiply that count by its corresponding face-surface area (yz, xz and xy, respectively) and sum up the result.
Errors are reduced form +/- 500% difference to +/- 15-30% difference, compared to matlab results. See also #117