BUG: Move cropping / pad resampling
Move cropping of the image to point after application of filter, before calculation of features.
Recode resampling, allowing padDistance. This enables to crop the image partly before applying a filter.

These changes are to prevent the LoG filter failing, as it needs a certain size in the image to function. If input image is too small, it throws an error. This size is needed to enable the LoG kernel to be passed around the image.
Default padding at resampling is set at 5 on all faces (in 3D), i.e. size of the resultant image will be bounding box + 2*5 in directions x, y and z. padDistance is defined as the number of voxels to pad in the new feature space.