ENH: Make application of filters dynamic
Functions that apply a filter now return a generator (which iterates over all input images generated by the filter) instead of a calculated feature vector.
This changes the workflow such, that first the selected filters are determined and stored in a generator, after which the features are calculated by iterating over this generator. Potential memory issues are minimised by making use of generators, which ensure that filtered images are calculated at the point where they are used for feature calculation.
Additionally, possible input image types are now determined dynamically, and are available from `getInputImageTypes`. This returns a list of type names, which enable the respective filters when present in `inputImage` as the key. By default, all filters are enabled.