ENH: Handle calculation of `shape` separately
Shape features are independent on the value of the grey level, and as such, different filters will not change the calculated values.
Therefore, handle shape calculation separately. If it should be calculated, calculate it once (on original image) and skip it in `calculateFeatures`, which is called for feature calculation on every inputImage (original and filtered).