ENH: Decrease computation time for GLCM
This is achieved by changing the iteration.
In the old version, iteration was done over all voxels in the segmentation, checking the entire neighbourhood of each voxel.
In this commit, iteration is now done over gray level twice (once as gray level i, once as j, corresponding to "center voxel" and "neighbour voxel").
Taking the intersection of indices to voxels that are neighbours to voxels with level i on the one hand and indices to voxels with gray level j on the other yields the indices to voxels that are BOTH have gray level j and are a neighbour to a voxel with gray level i.