ENH: Parameterize label
Allow user to set the value of the label in the mask image. This allows users to specify on which label the features should be calculated.
A standard label can be defined for a whole batch (kwargs['label']) or it can be passed to featureextractor.execute, in which case kwargs will be updated.
 By default, label = 1.