ENH: Add additional calls to logging
Add calls to logging throughout the package, with various levels and loggers specific to each module, so as to retain information on origin of logging call.
Add simple logging to file in `helloRadiomics.py` and switch from manual writing to file to using logging in `batchprocessing.py`.
Update comments in `__init__.py` and add them as the docstring of the `debug` function, so it is also listed in the documentation.
Initialize variable debugging (boolean, set by function `debug`) in `__init__.py`, so it is exposed throughout the package.