BUG: Revert to numpy functionality for image filters
Square, squareroot, logarithmic and exponential image filters.
Needed to better handle scaling and negative values:
Image intensities that are negative are filtered using its absolute value (if necessary), after which the resultant value is made negative again.
Additionally, in logarithm image filters, image intensities are scaled and offset in such a way, that there are no values between -1 and 1 (as the resultant value would go to -infinity)