ENH: Add optional weighting factor to GLCM, GLRLM.
Add the option the weight calculated matrices by a weighting factor, based upon the distance between two neighbouring voxels according to a specified weightingNorm.
If this is set, matrices are weighed and summed before feature calculation.
If not set, features are calculated on matrices for every angle separately, and the value returned is a mean for all angles.