ENH: Add parameter file parsing, update documentation of featureextractor
Add parameter file parser in imageoperation which can read a parameters file. This file can contain settings for feature(classes) to enable, which filters to apply and what settings should be set for preprocessing / feature extraction (settings as can be specified in kwargs).
This function returns three dictionaries, one for each type of settings. Detailed explanation on formatting of the parameters is supplied in the docstring, as well as in an example parameters file, which is added in batchExamples.
Add a function in `featureextractor.py` to use this parser.

Update documentation of featureextractor to reflect changes made in this and previous commits.
Additionally, handle enabling / disabling input images similar to enabling / disabling of feature classes.
Switch default enabled input images: only enable original input image by default (requires user to explicitly enable any extra filters).