ENH: Update algorithm for gldm and ngtdm
Original algorithm was very memory intensive (>4GB RAM in lung2), as it entailed creating a copy of the image for every angle.
Algorithm now only uses one copy of the image at a time. This copy is used to get a shifted array of the image, which can then be matched against the original image.