ENH: Add parameter file parsing, update documentation of featureextractor
Add parameter file parser in imageoperation which can read a parameters file. This file can contain settings for feature(classes) to enable, which filters to apply and what settings should be set for preprocessing / feature extraction (settings as can be specified in kwargs).
This function returns three dictionaries, one for each type of settings. Detailed explanation on formatting of the parameters is supplied in the docstring, as well as in an example parameters file, which is added in batchExamples.
Add a function in `featureextractor.py` to use this parser. And allow initialization of `featureextractor` using a parameters file.