ENH: Add command line tool
Takes an image and a corresponding mask and extracts features as specified by default settings or as specified in optional parameters file.
Output is written to stdout or specified outfile (mode: append). Default is one line for each feature with its calculated value. By specifying --format csv or json, output is formatted to a comma separated value list or a json file.