ENH: Add custom validation functions for interpolator, wavelet, weightingNorm
Custom validation is needed for interpolator and weightingNorm, as regular Enum does not support a None value, causing validation to fail when user attempts to set None. By using custom validation, a None value will be accepted here.
Additionally validate the wavelet setting against the list of possible values in pyWavelets, ensuring no errors arrise when pyWavelet does not have the current set, as well as enabling any newly added wavelets when they are added to pyWavelet. Update the documentation to reflect this.