BUG: Pad GLDZM
The distance image filter from simpleITK does not consider voxels on the edge of the image to be on the edge of a ROI (this enables a distancemap to be calculated in 2D). However, currently the mask image passed to the filter is cropped on the boundingbox, causing it to always have voxels on the border.
Therefore, pad the image and mask by 1 in those directions the size > 1 (ensuring a correct distancemap is also generated in case of 2D segmentations).