ENH: Store boundingBox after `cropToTumorMask`
Image is cropped after every filter applied, and separately for shape. Therefore, this would mean n+1 calls to `SimpleITK.LabelStatisticsImageFilter.Execute()`, where n is the number of filters applied. However, only one call should suffice, as the values of the bounding box aren't altered by the filters.
Therefore, store the value of the bounding box in `featureexctractor.execute()` and pass them to the `cropToTumorMask` on subsequent calls.
As the value of the bounding box is only stored inside the `execute` function, it will not cause a bounding box from one image to be used on another.

For the 5 testcases, this already improves computation time by 1+ seconds, where only 2 calls are made per image/mask combination (during use, this could increase easily to 20+ calls).