ENH: Change how baseline is stored
Store the baseline of each feature class in a separate file, this simplifies adding a feature class to the package.
Each baseline file is named as 'baseline_<className>.csv' and is a csv file containing the provenance information and the baseline values for the various features. Provenance information is used to dynamically determine the settings that need to be used during testing. The name of the feature class is extracted from the file name.

Updates tool `addClassToBaseline.py` to automatically generate correct files. When a new class is added to the code, this tool is run, after which the baseline using standard test settings (contained in this tool) is generated, no other actions are required to setup testing for this new class.
This reduces the steps needed to add a class to just defining the class and updating `docs\radiomics.rst` (to include the class in the automatic generation of documentation).

Additionally, move `custom_name_func` outside of class `RadiomicsTestUtils`, so it can also be used in `test_docstrings.py`, without having to instantiate `RadiomicsTestUtils`, which loads the entire baseline.