BUG: Fix error in C GLRLM
The inner  loop here runs from d2 = 1 to size[mDims[2]], meaning
it will run from x = 1 to size(x).
This is to prevent running over x = 0 (these start voxels are covered
in the previous loop, when x is fixed at 0).

However this applies to the case where angle[x] = 1. When angle[x] = -1,
x is fixed at size(x) in the previous loop and x has to run
from x = 0 to x = size(x) - 1. This should be handled by:
`if (angles[a * 3 + mDims[2]] < 0) jd[mDims[2]] --;`
However, mDims was set to mDims[1], meaning it applied the correction
to the wrong dimension. This caused the start voxels with x = 0 to be
erroneously omitted and start voxels with x = size(x) to be counted
twice. Additionally, when angly(y) is -1, the runs that would start
at y = 0 would be runs without elements, as it is erroneously set to
start at -1, which causes the while loop to exit immediately.