DOC: Update documentation
Add some additional references in GLRLM.
Fix some errors in documentation (old names, spelling errors).
Insert active links where documentation refers to other parts of the toolbox.
Add installation instructions for Windows.
Add documentation on some functions in imageoperations.

Update requirements.txt to use PyKwalify>=1.6.0, as this contains a fix for the bug which was breaking the output in jupyter. Remove hack around this bug in the helloRadiomics notebook example.

Move application of filters to `imageoperations` (instead of having two parts, one in `featureextractor`, one in `imageoperations`). Returns filtered image as part of a python generator. Apply changes to parts of pyradiomics that uses this functionality (mainly in `featureextractor.execute` and the helloFeatureClass examples).
N.B. This change causes the potential filters to have capitals in the name, similar to enabling individual features in a feature class. Changed the parameter schema accordingly.