ENH: Move getFeatureClasses to radiomics namespace
Move the `getFeatureClasses` and `getInputImages` out of the RadiomicsFeaturesExtractor class and to the radiomics namespace.
This makes these functions more globally available (as it is also used in the testing, which does not necessarily use the featureextractor).

Additionally, add new parameter, "additionalInfo" (Boolean) to control wheter or not to append the extra information on the extraction to the output. This additional information is provided by the generalinfo.py module.